Apply original class to the container
This patch alters the wrapping in such a way that any original classnames that are present on the target element, are also added to the `.wf-responsive-container` wrapper.

This makes it easier, especially when embedding iframes, to apply a certain ratio-class (square, 16:9, etc) in the editor, which is then also added to the container. The template css can then pick up that class and apply `height:0;padding-top:100%;` or whatever makes sense for the css.

I do suppose this might break some existing stuff relying on classnames on the iframe?? So maybe it should be put behind a toggle in the plugin settings?